Rework logging system (xlog.go) / Modify completion handling (WIP)
From now one, logrus is to be completely contained w/in xlog.go and xlog
becomes the replacement logging api.

Various line length and other cleanups.

EP completion handling being redesigned so that all fnotify events are
handled in the same place regardless of the operation type
(SystemInfoOp, IdentifyOp, etc.)

Stuff is still massively broken as of this patch..